# doc/api-doc.yml
openapi: 3.0.0
info:
  title: Melodify
  description: API for a music-based social media application
  version: 0.1.0
servers:
  - url: 'http://34.118.44.165/api'
    description: Main API server
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - surname
                - email
                - username
                - password
              properties:
                name:
                  type: string
                surname:
                  type: string
                email:
                  type: string
                username:
                  type: string
                password:
                  type: string
            example:
              name: "John"
              surname: "Doe"
              email: "john_doe@gmail.com"
              username: "john_doe"
              password: "password123"  
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  message:
                    type: string
              example:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                message: "Registration successful"
        '400':
          description: Bad request
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
            example:
              username: "john_doe"
              password: "password123"
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  message:
                    type: string
              example:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                message: "Login successful"
        '401':
          description: Unauthorized
...